Fixes issue #1.
It looks like this line in the original `get` definition

```
 if (!this.Store[key]) return;
```

was designed to avoid `SyntaxError`s when trying to clone `undefined`. Since this is really a concern for `clone`, not `get`, I propose moving the check into the `clone` definition. 

Furthermore, `null` is a legitimate value in JSON, so I propose we just check for `undefined`.